@@ -9,19 +9,19 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "17.3.7",
-    "@angular/cdk": "17.3.7",
-    "@angular/common": "17.3.7",
-    "@angular/compiler": "17.3.7",
-    "@angular/core": "17.3.7",
-    "@angular/forms": "17.3.7",
-    "@angular/material": "17.3.7",
-    "@angular/platform-browser": "17.3.7",
-    "@angular/platform-browser-dynamic": "17.3.7",
-    "@angular/router": "17.3.7",
+    "@angular/animations": "18.0.0",
+    "@angular/cdk": "18.0.0",
+    "@angular/common": "18.0.0",
+    "@angular/compiler": "18.0.0",
+    "@angular/core": "18.0.0",
+    "@angular/forms": "18.0.0",
+    "@angular/material": "18.0.0",
+    "@angular/platform-browser": "18.0.0",
+    "@angular/platform-browser-dynamic": "18.0.0",
+    "@angular/router": "18.0.0",
     "@ctrl/tinycolor": "4.1.0",
-    "@ng-matero/extensions": "17.3.0",
-    "chart.js": "4.4.2",
+    "@ng-matero/extensions": "17.3.2",
+    "chart.js": "4.4.3",
     "chartjs-adapter-date-fns": "3.0.0",
     "chartjs-plugin-annotation": "3.0.1",
     "date-fns": "3.6.0",
@@ -29,12 +29,12 @@
     "ngx-color": "9.0.0",
     "rxjs": "7.8.1",
     "tslib": "2.6.2",
-    "zone.js": "0.14.5"
+    "zone.js": "0.14.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.6",
-    "@angular/cli": "17.3.6",
-    "@angular/compiler-cli": "17.3.7",
+    "@angular/build": "^18.0.1",
+    "@angular/cli": "18.0.1",
+    "@angular/compiler-cli": "18.0.0",
     "typescript": "5.4.5"
   }
 }
