@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule } from '@angular/material/dialog';
@@ -13,26 +13,19 @@ import { ChartComponent } from './chart.component';
 
 import { ChartService } from '../services/chart.service';
 
-@NgModule({
-  declarations: [
-    ChartComponent
-  ],
-  imports: [
-    CommonModule,
-    HttpClientModule,
-
-    MatButtonModule,
-    MatDialogModule,
-    MatIconModule,
-    MatToolbarModule,
-    MatTooltipModule
-  ],
-  exports: [
-    ChartComponent
-  ],
-  providers: [
-    ChartService
-  ],
-  bootstrap: [ChartComponent]
-})
+@NgModule({ declarations: [
+        ChartComponent
+    ],
+    exports: [
+        ChartComponent
+    ],
+    bootstrap: [ChartComponent], imports: [CommonModule,
+        MatButtonModule,
+        MatDialogModule,
+        MatIconModule,
+        MatToolbarModule,
+        MatTooltipModule], providers: [
+        ChartService,
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class ChartModule { }
